Reorganize code in spectrum-ergo-connector
