+++ spectrum-ergo-connector/src/lib.rs
@@ -5,9 +5,13 @@ pub mod committee;
 pub mod deposit;
 pub mod rocksdb;
 pub mod script;
-pub mod vault;
+pub mod tx_event;
+pub mod tx_in_progress;
+pub mod vault_handler;
+pub mod vault_utxo;
 
 #[derive(Deserialize, Serialize, Debug, PartialEq, Eq, Clone)]
+/// Ergo-specific information to send to the consensus-driver.
 pub struct AncillaryVaultInfo {
     pub tx_id: TxId,
     pub box_id: BoxId,
