+++ spectrum-ergo-connector/src/rocksdb.rs
@@ -1,4 +1,4 @@
-pub mod deposits;
+pub mod deposit;
 pub mod ergo_tx_event_history;
 pub mod tx_retry_scheduler;
 pub mod vault_boxes;
