+++ spectrum-ergo-connector/src/rocksdb/deposit.rs
@@ -4,23 +4,8 @@ use async_std::task::spawn_blocking;
 use async_trait::async_trait;
 use ergo_lib::ergotree_ir::chain::ergo_box::BoxId;
 use rocksdb::{Direction, IteratorMode, ReadOptions};
-use serde::{Deserialize, Serialize};
-use spectrum_chain_connector::InboundValue;
-use spectrum_offchain_lm::data::AsBox;
 
-use crate::script::ErgoInboundCell;
-
-#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
-pub struct ProcessedDeposit(pub AsBox<ErgoInboundCell>);
-
-#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
-pub struct UnprocessedDeposit(pub AsBox<ErgoInboundCell>);
-
-impl From<UnprocessedDeposit> for InboundValue<BoxId> {
-    fn from(value: UnprocessedDeposit) -> Self {
-        InboundValue::from(value.0 .1)
-    }
-}
+use crate::deposit::{ProcessedDeposit, UnprocessedDeposit};
 
 #[async_trait(?Send)]
 pub trait DepositRepo {
@@ -155,7 +140,7 @@ fn prefixed_key(prefix: &str, box_id: &BoxId) -> Vec<u8> {
 
 #[cfg(test)]
 mod tests {
-    use crate::rocksdb::deposits::{DepositRepo, DepositRepoRocksDB, ProcessedDeposit, UnprocessedDeposit};
+    use crate::rocksdb::deposit::{DepositRepo, DepositRepoRocksDB, ProcessedDeposit, UnprocessedDeposit};
     use crate::script::tests::gen_random_token;
     use crate::script::{ErgoCell, ErgoInboundCell};
     use blake2::digest::crypto_common::rand_core::RngCore;
