+++ spectrum-ergo-connector/src/rocksdb/ergo_tx_event_history.rs
@@ -2,22 +2,10 @@ use std::sync::Arc;
 
 use async_std::task::spawn_blocking;
 use async_trait::async_trait;
-use ergo_lib::chain::transaction::TxId;
-use ergo_lib::ergotree_ir::chain::ergo_box::BoxId;
 use log::info;
 use rocksdb::{Direction, IteratorMode};
-use serde::{Deserialize, Serialize};
-use spectrum_chain_connector::{ChainTxEvent, InboundValue, SpectrumTx, SpectrumTxType, VaultBalance};
-use spectrum_ledger::{
-    cell::{ProgressPoint, SValue, TermCell},
-    interop::Point,
-    ChainId,
-};
 
-use crate::script::{ErgoInboundCell, ErgoTermCell};
-use crate::AncillaryVaultInfo;
-
-use super::vault_boxes::VaultUtxo;
+use crate::tx_event::ErgoTxEvent;
 
 /// Store the entire history of `ErgoTxEvents`, allowing a new consensus-driver to sync with
 /// the ergo-connector.
@@ -26,124 +14,6 @@ pub trait ErgoTxEventHistory {
     async fn append(&mut self, moved_value: ErgoTxEvent);
     /// Returns `ErgoTxEvent` that is closest and >= `height`.
     async fn get(&self, height: u32) -> Option<(ErgoTxEvent, u32)>;
-}
-
-#[derive(Deserialize, Serialize, Debug, PartialEq, Eq, Clone)]
-pub struct SpectrumErgoTx {
-    pub progress_point: u32,
-    pub tx_id: TxId,
-    pub tx_type: ErgoTxType,
-}
-
-#[derive(Deserialize, Serialize, Debug, PartialEq, Eq, Clone)]
-pub enum ErgoTxType {
-    /// Spectrum Network deposit transaction
-    Deposit {
-        /// Value that is inbound to Spectrum-network
-        imported_value: Vec<ErgoInboundCell>,
-        vault_info: (VaultUtxo, AncillaryVaultInfo),
-    },
-
-    /// Spectrum Network withdrawal transaction
-    Withdrawal {
-        /// Value that was successfully exported from Spectrum-network to some recipient on-chain.
-        exported_value: Vec<ErgoTermCell>,
-        vault_info: (VaultUtxo, AncillaryVaultInfo),
-    },
-
-    NewUnprocessedDeposit(ErgoInboundCell),
-    RefundedDeposit(ErgoInboundCell),
-}
-
-impl From<SpectrumErgoTx> for SpectrumTx<BoxId, AncillaryVaultInfo> {
-    fn from(value: SpectrumErgoTx) -> Self {
-        let SpectrumErgoTx {
-            progress_point,
-            tx_type,
-            ..
-        } = value;
-        let progress_point = ProgressPoint {
-            chain_id: ChainId::from(0),
-            point: Point::from(progress_point as u64),
-        };
-        match tx_type {
-            ErgoTxType::Deposit {
-                imported_value,
-                vault_info: (vault_utxo, ancillary_info),
-            } => {
-                let imported_value = imported_value.into_iter().map(InboundValue::from).collect();
-
-                let vault_balance = VaultBalance {
-                    value: SValue::from(&vault_utxo),
-                    on_chain_characteristics: ancillary_info,
-                };
-                SpectrumTx {
-                    progress_point,
-                    tx_type: SpectrumTxType::Deposit {
-                        imported_value,
-                        vault_balance,
-                    },
-                }
-            }
-            ErgoTxType::Withdrawal {
-                exported_value,
-                vault_info: (vault_utxo, ancillary_info),
-            } => {
-                let exported_value = exported_value.into_iter().map(TermCell::from).collect();
-                let vault_balance = VaultBalance {
-                    value: SValue::from(&vault_utxo),
-                    on_chain_characteristics: ancillary_info,
-                };
-                SpectrumTx {
-                    progress_point,
-                    tx_type: SpectrumTxType::Withdrawal {
-                        exported_value,
-                        vault_balance,
-                    },
-                }
-            }
-            ErgoTxType::NewUnprocessedDeposit(inbound_cell) => {
-                let inbound_value = InboundValue::from(inbound_cell);
-                SpectrumTx {
-                    progress_point,
-                    tx_type: SpectrumTxType::NewUnprocessedDeposit(inbound_value),
-                }
-            }
-
-            ErgoTxType::RefundedDeposit(inbound_cell) => {
-                let inbound_value = InboundValue::from(inbound_cell);
-                SpectrumTx {
-                    progress_point,
-                    tx_type: SpectrumTxType::RefundedDeposit(inbound_value),
-                }
-            }
-        }
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug, PartialEq, Eq, Clone)]
-pub enum ErgoTxEvent {
-    /// A new set of TXs are made on-chain for a given progress point.
-    Applied(SpectrumErgoTx),
-    /// When the chain experiences a rollback, movements of value must be unapplied.
-    Unapplied(SpectrumErgoTx),
-}
-
-impl ErgoTxEvent {
-    pub fn get_height(&self) -> u32 {
-        match self {
-            ErgoTxEvent::Applied(tx) | ErgoTxEvent::Unapplied(tx) => tx.progress_point,
-        }
-    }
-}
-
-impl From<ErgoTxEvent> for ChainTxEvent<BoxId, AncillaryVaultInfo> {
-    fn from(value: ErgoTxEvent) -> Self {
-        match value {
-            ErgoTxEvent::Applied(tx) => ChainTxEvent::Applied(SpectrumTx::from(tx)),
-            ErgoTxEvent::Unapplied(tx) => ChainTxEvent::Unapplied(SpectrumTx::from(tx)),
-        }
-    }
 }
 
 pub struct ErgoTxEventHistoryRocksDB {
@@ -258,15 +128,14 @@ mod tests {
     use sigma_test_util::force_any_val;
 
     use crate::{
-        rocksdb::{
-            ergo_tx_event_history::{ErgoTxEventHistory, InMemoryMovedValueHistory},
-            vault_boxes::VaultUtxo,
-        },
+        rocksdb::ergo_tx_event_history::{ErgoTxEventHistory, InMemoryMovedValueHistory},
         script::{ErgoCell, ErgoTermCell},
+        tx_event::{ErgoTxEvent, ErgoTxType, SpectrumErgoTx},
+        vault_utxo::VaultUtxo,
         AncillaryVaultInfo,
     };
 
-    use super::{ErgoTxEvent, ErgoTxEventHistoryRocksDB, ErgoTxType, SpectrumErgoTx};
+    use super::ErgoTxEventHistoryRocksDB;
 
     #[tokio::test]
     async fn test_rocksdb_single_insertion() {
